+++ src/main/java/com/project/planb/common/security/details/PrincipalDetails.java
@@ -1,4 +1,4 @@
-package com.project.planb.common.security;
+package com.project.planb.common.security.details;
 
 import com.project.planb.domain.member.entity.Member;
 import lombok.Getter;
