+++ src/main/java/com/project/planb/domain/member/controller/MemberController.java
@@ -1,8 +1,8 @@
 package com.project.planb.domain.member.controller;
 
+import com.project.planb.common.security.dto.TokenResDto;
 import com.project.planb.domain.member.dto.MemberJoinReqDto;
 import com.project.planb.domain.member.dto.MemberLoginReqDto;
-import com.project.planb.common.security.dto.TokenResDto;
 import com.project.planb.domain.member.service.MemberService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
