+++ src/main/java/com/project/planb/domain/member/service/MemberService.java
@@ -1,12 +1,12 @@
 package com.project.planb.domain.member.service;
 
-import com.project.planb.domain.member.dto.MemberJoinReqDto;
-import com.project.planb.domain.member.dto.MemberLoginReqDto;
-import com.project.planb.common.security.dto.TokenResDto;
-import com.project.planb.domain.member.entity.Member;
 import com.project.planb.common.exception.CustomException;
 import com.project.planb.common.exception.ErrorCode;
+import com.project.planb.common.security.dto.TokenResDto;
 import com.project.planb.common.security.jwt.JwtTokenProvider;
+import com.project.planb.domain.member.dto.MemberJoinReqDto;
+import com.project.planb.domain.member.dto.MemberLoginReqDto;
+import com.project.planb.domain.member.entity.Member;
 import com.project.planb.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
